AWBW: Changed update mechanism & default shortcut
The gist version didn't really work out because there didn't seem to be
any way to link to the latest version of the script at any given time,
thus completely messing up the monkey extensions' update mechanisms.

Hardcoding it seems to do the job.

Also, the default shortcut was changed because O is generally a bad
shortcut to have since I personally find it hard to tell it apart from 0
(that's a zero). K stands for Kustomize or something.